DynamicAddUsers: Automatically move conflicting old accounts' emails out of the way.
WordPress only allows a single account to be registered with a given email
address. Because Middlebury email addresses were reused for many years, there
are cases where a new user gets an email address that is already associated
with another user in the WordPress database.

In these cases, confirm that the user is no longer in the directory and then
update the old email address to add the `-old-replaced` suffix and then try
adding the new account again.