Use the plugin->getUserManager() method to access the user manager when needed.
For #1. By waiting until the user manager is actually needed we can avoid
loading configuration for the directory service it uses for most requests,
only fully configuring it when it will be utilized.

The implementation in 5448759cc0cb3a94ff389e579a3044ee5dee4e22 will operate
fine, but when the AddUsers class in instantiated every page load (in order
to trigger its hooks to add to menus and add actions) it will also
end up needing to look up which Directory implementation is configured and
load that service. By accessing it when needed a bare minimum of
configuration can be loaded on most requests.